osd: separate hb front client messenger for pre-luminous
The legacy hb messengers set the cluster protocol to CEPH_OSD_PROTOCOL,
but the normal OSD front messenger (which we are now pinging instead)
does not.  Use a separate messenger for that.

This is all temporary; post-luminous we'll rip out the legacy messengers
since we won't have to talk to pre-luminous OSDs.

Signed-off-by: Sage Weil <sage@redhat.com>